Combat - Split handle function into two functions
--WHAT--
Split the "handleChosenOption" function into two:
1) carryOutAction
2) determineOptionAndPrint

--WHY--
The handleChosenOption function would modify the player and enemy
objects and then print the status. I believe this was the cause of
the bug that would prevent the output from updating when it should.

Regardless, now that the function is split, the program is more
modular. Which is a good thing and an improvement.