Temp Commit - See notes for where to pick up from
--WHAT--
I'm trying to process blocks first so that, when user or enemy attacks,
its already known if the attack will be blocked or not.

--HOW--
I need to rethink how to modularize this action. New functions will
have to be made (check out the phase1 and phase2 functions) and old
functions will have to be modified.