Combat/TestStuff.java - Make output more modular
--TestStuff.java--
Add test to see what happens
when strings our multiline

--Combat.java - PrintScreen--
Revise so that each line is a string,
making the code more modular and easier
to revise and read.

--Combat.java - TwoPlayerMode--
Revise so that the code that handles this
is a single function that calls other
functions, thus making the code more modular
and readable, and maintainable.