Combat/gitignore - Add animate function, add swo
--GITIGNORE--
Add swo files to file that should be ignored.
These are files that let the computer know that another file in vim
is open and should not be edited by the second editor.

--COMBAT-Animate--
Added a moveCursor function that moves the cursor x spaces to the
left/right/down/up.

Add a getSpaces function that returns a string with the specified
number of spaces.

Start making an animate function that will create animation.